# contracts/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ICGL Sovereign Cockpit API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/agents/{agent_id}/history": {
      "get": {
        "tags": [
          "Agents Portfolio"
        ],
        "summary": "Get Agent History",
        "description": "Returns the work history of a specific agent.",
        "operationId": "get_agent_history_api_agents__agent_id__history_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/stats": {
      "get": {
        "tags": [
          "Agents Portfolio"
        ],
        "summary": "Get Agent Stats",
        "description": "Returns calculated analytics for the agent.",
        "operationId": "get_agent_stats_api_agents__agent_id__stats_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_id}/role": {
      "get": {
        "tags": [
          "Agents Portfolio"
        ],
        "summary": "Get Agent Role",
        "description": "Returns the dynamic role definition (System Prompt) from the live agent instance.",
        "operationId": "get_agent_role_api_agents__agent_id__role_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/executive/sign/{item_id}": {
      "post": {
        "tags": [
          "Executive Agent"
        ],
        "summary": "Sign Off Action",
        "description": "Sovereign Human Sign-off.\nThis authorizes the action for execution (in a future step).",
        "operationId": "sign_off_action_api_executive_sign__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executive/reject/{item_id}": {
      "post": {
        "tags": [
          "Executive Agent"
        ],
        "summary": "Reject Action",
        "description": "Sovereign Human Rejection.\nCancels the action.",
        "operationId": "reject_action_api_executive_reject__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/governance/decisions/register": {
      "post": {
        "summary": "Register Decision Api",
        "description": "Registers a human decision for an ADR and persists it.",
        "operationId": "register_decision_api_api_governance_decisions_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/governance/conflicts": {
      "get": {
        "summary": "Get Conflicts",
        "description": "Return stored conflicts from persistent registry.",
        "operationId": "get_conflicts_api_governance_conflicts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Conflict",
        "description": "Create a conflict entry persistently.",
        "operationId": "create_conflict_api_governance_conflicts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Conflict"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/governance/proposals/{proposal_id}": {
      "patch": {
        "summary": "Patch Proposal Api",
        "description": "Basic proposal patching: currently supports status update only.",
        "operationId": "patch_proposal_api_api_governance_proposals__proposal_id__patch",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sign/{adr_id}": {
      "post": {
        "summary": "Sign Decision",
        "operationId": "sign_decision_sign__adr_id__post",
        "parameters": [
          {
            "name": "adr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adr Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionRegisterRequest": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "signed_by": {
            "type": "string",
            "title": "Signed By",
            "default": "operator"
          }
        },
        "type": "object",
        "required": [
          "proposal_id",
          "decision",
          "rationale"
        ],
        "title": "DecisionRegisterRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "human_id": {
            "type": "string",
            "title": "Human Id",
            "default": "bakheet"
          }
        },
        "type": "object",
        "required": [
          "action",
          "rationale"
        ],
        "title": "SignRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}